pyproject.toml
[project]
name = "agentic-app"
version = "0.1.0"
description = "Music recommendation agent dyn'tion app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.19.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.0.0",
    "google-genai>=1.46.0",
]
